ft: normal estimation
