+++ src/utils/pointcloud_io.cpp
@@ -42,39 +42,6 @@ std::optional<PC_NORM::Ptr> loadPcNormals(ILogger& logger, const std::string& pc
 }
 /*//}*/
 
-/*//{ hasNormals() */
-bool hasNormals(ILogger& logger, const std::string& pcd_file)
-{
-  pcl::PCDReader reader_pcd;
-  pcl::PCLPointCloud2 pc;
-  if (reader_pcd.readHeader(pcd_file, pc) < 0)
-  {
-    logger.error("[PCLSupportLibrary] Couldn't read header of PCD file: " + pcd_file);
-    return false;
-  }
-
-  unsigned int normal_fields = 0;
-  bool curvature_field = false;
-  for (const auto& field : pc.fields)
-  {
-    if (field.name.rfind("normal", 0) == 0)
-    {
-      normal_fields++;
-    }
-    else if (field.name == "curvature")
-    {
-      curvature_field = true;
-    }
-  }
-  if (normal_fields != 3 || !curvature_field)
-  {
-    return false;
-  }
-
-  return true;
-}
-/*//}*/
-
 /*//{ savePCD() */
 void savePCD(const std::string& pcd_file, const pcl::PCLPointCloud2& cloud, const bool binary)
 {
